+++ src/docs/components/slides/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+export { default as ParallaxSlide } from './parallaxSlideDoc';
